Strict typescript front and back  [clil1pcgc0002t40krdqam23d]

+++ src/components/Questions/QuestionsContent/FlashcardDialog/StartedFlashcardDialogChild/QuestionDialog/useDefaultDeleteQuestion/useDefaultDeleteQuestion.ts
@@ -20,6 +20,9 @@ export const useDefaultDeleteQuestion = () => {
       title: "Delete question?",
       description: "Are you sure you want to delete this question?",
       onConfirm: () => {
+        if (!questionId) {
+          return
+        }
         axios.delete(urls.api.define.questionId(questionId)).then(() => {
           setAllQuestions(allQuestions.filter((q) => q.id !== questionId))
           onClose()
